Seperate inventory interactions The goal is to improve code readability and compile times
