+++ CMakeLists.txt
@@ -80,6 +80,10 @@ set(SHARED_SOURCES
     src/bpp_shared/world/storage/region.cpp
     src/bpp_shared/world/storage/region_manager.cpp
     src/bpp_shared/blocks/block_properties.cpp
+    src/bpp_shared/inventory/inventory_interaction.cpp
+    src/bpp_shared/inventory/interactions/chest.cpp
+    src/bpp_shared/inventory/interactions/large_chest.cpp
+    src/bpp_shared/inventory/interactions/player.cpp
     src/bpp_shared/crafting/recipe_manager.cpp
     src/bpp_shared/crafting/vanilla_recipes.cpp
