+++ src/bpp_server/commands/implementation/command_give.cpp
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: AGPL-3.0-only
 */
 
-#include "../../packet/handle_packet.h"
+#include "../../packet/packet_utils.h"
 #include "../command.h"
 #include "../command_manager.h"
 #include "inventory/item_stack.h"
@@ -15,8 +15,10 @@
 // Give yourself a block or item
 // Usage:
 //   /give <id>[:meta] [amount]
-std::string CommandGive::Execute(std::vector<std::string>& parameters, PlayerSession& session, [[maybe_unused]] WorldManager& world,
-                                 [[maybe_unused]] std::function<void(PlayerSession&)> transferDimension, [[maybe_unused]] Server& server) {
+std::string CommandGive::Execute(std::vector<std::string>& parameters, PlayerSession& session,
+                                 [[maybe_unused]] WorldManager& world,
+                                 [[maybe_unused]] std::function<void(PlayerSession&)> transferDimension,
+                                 [[maybe_unused]] Server& server) {
 	// TODO: Let player specify another player to give to
 	if (parameters.size() <= 1)
