+++ src/bpp_shared/inventory/inventory_interaction.h
@@ -1,14 +1,12 @@
 /*
  * Copyright (c) 2026, Aidan <JcbbcEnjoyer>
+ * Copyright (c) 2026, jwaxy <jwaxy.is-a.dev>
  *
  * SPDX-License-Identifier: AGPL-3.0-only
 */
 #pragma once
-#include "inventories.h"
+#include "inventory.h"
 #include "inventory/item_stack.h"
-#include "items.h"
-#include "tile_entities/tile_entity.h"
-#include <memory>
 
 struct DeltaSlot {
 	ItemStack stack;
@@ -21,378 +19,21 @@ struct InventoryInteraction {
 	ItemStack carried;
 	Inventory* inventory;
 
-	InventoryInteraction(Inventory* inv) : inventory(inv) {}
+	InventoryInteraction(Inventory* inv);
 	virtual ~InventoryInteraction() = default;
 
-	virtual bool canExist() {
-		return inventory != nullptr;
-	}
+	virtual bool canExist();
 
 	// Take a snapshot of this inventory
-	virtual void initSnapshot() {
-		snapshot = inventory->slots;
-	}
+	virtual void initSnapshot();
 
 	// Analyze the snapshot vs the current inventory
 	// Returns a list of slots that are different
-	virtual std::vector<DeltaSlot> tickDiff() {
-		std::vector<DeltaSlot> differences;
-		for (size_t i = 0; i < snapshot.size(); i++) {
-			[[maybe_unused]] auto* current = inventory->getStackInSlot(i);
-			auto& snap = snapshot[i];
+	virtual std::vector<DeltaSlot> tickDiff();
 
-			bool changed = snap != inventory->slots[i];
-			if (!changed)
-				continue;
+	virtual void onLeftClick(int slot);
 
-			snap = inventory->slots[i];
-			differences.push_back({ snap, int(i) });
-		}
-		return differences;
-	}
+	virtual void onRightClick(int slot);
 
-	virtual void onLeftClick(int slot) {
-		auto targetSlot = inventory->getStackInSlot(slot);
-
-		// Empty slot
-		if (!targetSlot) {
-			if (carried.id != ITEM_INVALID) {
-				inventory->setInventorySlotContents(slot, &carried);
-				carried = ItemStack{};
-			}
-			inventory->onInventoryChanged();
-			return;
-		}
-
-		// Not carrying anything
-		if (carried.id == ITEM_INVALID) {
-			carried = *targetSlot;
-			inventory->clearSlot(slot);
-			inventory->onInventoryChanged();
-			return;
-		}
-
-		// Same item; merge
-		if (targetSlot->id == carried.id && targetSlot->data == carried.data) {
-			int maxStack = GetMaxStack(targetSlot->id);
-			int space = maxStack - targetSlot->count;
-			int toMove = std::min(space, (int)carried.count);
-			targetSlot->count += toMove;
-			carried.count -= toMove;
-			if (carried.count == 0)
-				carried = ItemStack{};
-			inventory->onInventoryChanged();
-			return;
-		}
-
-		// Different item; swap
-		ItemStack temp = *targetSlot;
-		*targetSlot = carried;
-		carried = temp;
-		inventory->onInventoryChanged();
-	}
-
-	virtual void onRightClick(int slot) {
-		auto targetSlot = inventory->getStackInSlot(slot);
-
-		if (carried.id != ITEM_INVALID) {
-			if (!targetSlot) {
-				ItemStack single{ carried.id, 1, carried.data };
-				inventory->setInventorySlotContents(slot, &single);
-				carried.count -= 1;
-				if (carried.count == 0)
-					carried = ItemStack{};
-				inventory->onInventoryChanged();
-				return;
-			}
-
-			// If we right click on the same item we are carrying just add one
-			if (targetSlot->id == carried.id && targetSlot->data == carried.data) {
-				int maxStack = GetMaxStack(targetSlot->id);
-				int space = maxStack - targetSlot->count;
-				if (space >= 1) {
-					targetSlot->count += 1;
-					carried.count -= 1;
-					if (carried.count == 0)
-						carried = ItemStack{};
-					inventory->onInventoryChanged();
-				}
-				return;
-			}
-
-			// If we right click on a different item, swap the cursor and that item
-			ItemStack temp = *targetSlot;
-			*targetSlot = carried;
-			carried = temp;
-			inventory->onInventoryChanged();
-			return;
-		}
-
-		if (!targetSlot)
-			return;
-
-		// Only split items if there stack count is greater than 1 and we aren't carrying anything
-		if (targetSlot->count > 1) {
-			// Beta always take the higher of the two if uneven
-			int taken = (targetSlot->count + 1) / 2;
-			int left = targetSlot->count - taken;
-			targetSlot->count = int8_t(left);
-			carried = ItemStack{ targetSlot->id, int8_t(taken), targetSlot->data };
-			inventory->onInventoryChanged();
-			return;
-		}
-
-		// If its only one item we just pick it up
-		carried = *targetSlot;
-		inventory->clearSlot(slot);
-		inventory->onInventoryChanged();
-		return;
-	}
-
-	virtual void onShiftClick(int slot) {
-		auto targetSlot = inventory->getStackInSlot(slot);
-		if (!targetSlot)
-			return;
-		inventory->mergeItemStackInInventory(*targetSlot);
-	}
-};
-
-struct LargeChestInventoryInteraction : InventoryInteraction {
-	InventoryPlayer* playerInventory;
-	std::weak_ptr<TileEntityChest> upperChest;
-	std::weak_ptr<TileEntityChest> lowerChest;
-	InventoryLargeChest chestInventory;
-
-	struct SharedInventory : Inventory {
-		LargeChestInventoryInteraction* owner = nullptr;
-		SharedInventory() : Inventory(90) {}
-		void onInventoryChanged() override {
-			if (owner)
-				owner->writeBack();
-		}
-	} sharedInventory;
-
-	LargeChestInventoryInteraction(InventoryPlayer* pinv, std::shared_ptr<TileEntityChest> upper,
-	                               std::shared_ptr<TileEntityChest> lower)
-	    : InventoryInteraction(&sharedInventory), playerInventory(pinv), upperChest(upper), lowerChest(lower),
-	      chestInventory(&upper->inventory, &lower->inventory) {
-		sharedInventory.owner = this;
-		mergeInventories();
-	}
-
-	virtual ~LargeChestInventoryInteraction() {
-		if (canExist())
-			writeBack();
-	}
-
-	virtual bool canExist() override {
-		return !upperChest.expired() && !lowerChest.expired();
-	}
-
-	// Only snapshot the chest portion we don't need the attached inventory
-	void initSnapshot() override {
-		snapshot.resize(size_t(chestInventory.getSizeInventory()));
-		for (size_t i = 0; i < size_t(chestInventory.getSizeInventory()); i++) {
-			auto* stack = chestInventory.getStackInSlot(i);
-			snapshot[i] = stack ? *stack : ItemStack{};
-		}
-	}
-
-	// Analyze the snapshot vs the current chest inventory
-	std::vector<DeltaSlot> tickDiff() override {
-		std::vector<DeltaSlot> differences;
-		for (size_t i = 0; i < snapshot.size(); i++) {
-			auto* currentPtr = chestInventory.getStackInSlot(int(i));
-			auto current = currentPtr ? *currentPtr : ItemStack{};
-
-			if (snapshot[i] == current)
-				continue;
-
-			snapshot[i] = current;
-			differences.push_back({ snapshot[i], int(i) });
-		}
-		mergeInventories();
-		return differences;
-	}
-
-	void mergeInventories() {
-		size_t slotCount = 0;
-		for (size_t i = 0; i < size_t(chestInventory.getSizeInventory()); i++) {
-			auto* stack = chestInventory.getStackInSlot(i);
-			sharedInventory.slots[slotCount++] = stack ? *stack : ItemStack{};
-		}
-		for (size_t i = 9; i <= 44; i++)
-			sharedInventory.slots[slotCount++] = playerInventory->slots[i];
-	}
-
-	void writeBack() {
-		for (size_t i = 0; i < 54; i++) {
-			auto& slot = sharedInventory.slots[i];
-			ItemStack* ptr = slot.id != ITEM_INVALID ? &slot : nullptr;
-			chestInventory.setInventorySlotContents(i, ptr);
-		}
-		for (size_t i = 54; i < 90; i++)
-			playerInventory->slots[i - 54 + 9] = sharedInventory.slots[i];
-	}
-
-	void onShiftClick(int slot) override {
-		auto stack = sharedInventory.getStackInSlot(slot);
-		if (!stack)
-			return;
-
-		ItemStack copy = *stack;
-
-		if (slot <= 53) {
-			// Chest -> inventory
-			[[maybe_unused]] bool success = playerInventory->mergeItemStackInInventory(copy, true, 9, 44);
-		} else {
-			// Inventory -> Chest
-			[[maybe_unused]] bool success = chestInventory.mergeItemStackInInventory(copy);
-		}
-
-		// Update the source in the real inventory before re-merging
-		if (slot <= 53) {
-			ItemStack* ptr = copy.count == 0 ? nullptr : &copy;
-			chestInventory.setInventorySlotContents(slot, ptr);
-		} else {
-			int playerSlot = slot - 54 + 9;
-			playerInventory->slots[size_t(playerSlot)] = copy.count == 0 ? ItemStack{} : copy;
-		}
-
-		// Re-sync sharedInventory from the real inventories
-		mergeInventories();
-	}
-};
-
-struct ChestInventoryInteraction : InventoryInteraction {
-	InventoryPlayer* playerInventory;
-	std::weak_ptr<TileEntityChest> chestHandle;
-	InventoryChest* chestInventory;
-
-	struct SharedInventory : Inventory {
-		ChestInventoryInteraction* owner = nullptr;
-		SharedInventory() : Inventory(63) {}
-		void onInventoryChanged() override {
-			if (owner)
-				owner->writeBack();
-		}
-	} sharedInventory;
-
-	ChestInventoryInteraction(InventoryPlayer* pinv, std::shared_ptr<TileEntityChest> chest)
-	    : InventoryInteraction(&sharedInventory), playerInventory(pinv), chestHandle(chest),
-	      chestInventory(&chest->inventory) {
-		sharedInventory.owner = this;
-		mergeInventories();
-	}
-
-	virtual ~ChestInventoryInteraction() {
-		if (canExist())
-			writeBack();
-	}
-
-	virtual bool canExist() override {
-		return !chestHandle.expired();
-	}
-
-	// Only snapshot the chest portion we don't need the attached inventory
-	void initSnapshot() override {
-		snapshot.resize(size_t(chestInventory->getSizeInventory()));
-		for (size_t i = 0; i < size_t(chestInventory->getSizeInventory()); i++)
-			snapshot[i] = chestInventory->slots[i];
-	}
-
-	// Analyze the snapshot vs the current chest inventory
-	std::vector<DeltaSlot> tickDiff() override {
-		std::vector<DeltaSlot> differences;
-		for (size_t i = 0; i < snapshot.size(); i++) {
-			[[maybe_unused]] auto* current = chestInventory->getStackInSlot(i);
-			auto& snap = snapshot[i];
-
-			bool changed = snap != chestInventory->slots[i];
-			if (!changed)
-				continue;
-
-			snap = chestInventory->slots[i];
-			differences.push_back({ snap, int(i) });
-		}
-		mergeInventories();
-		return differences;
-	}
-
-	void mergeInventories() {
-		size_t slotCount = 0;
-		for (auto& slot : chestInventory->slots)
-			sharedInventory.slots[slotCount++] = slot;
-		for (size_t i = 9; i <= 44; i++)
-			sharedInventory.slots[slotCount++] = playerInventory->slots[i];
-	}
-
-	void writeBack() {
-		for (size_t i = 0; i < 27; i++)
-			chestInventory->slots[i] = sharedInventory.slots[i];
-		for (size_t i = 27; i < 63; i++)
-			playerInventory->slots[i - 27 + 9] = sharedInventory.slots[i];
-	}
-
-	void onShiftClick(int slot) override {
-		auto stack = sharedInventory.getStackInSlot(slot);
-		if (!stack)
-			return;
-
-		ItemStack copy = *stack;
-
-		if (slot <= 26) {
-			// Chest -> inventory
-			[[maybe_unused]] bool success = playerInventory->mergeItemStackInInventory(copy, true, 9, 44);
-		} else {
-			// Inventory -> Chest
-			[[maybe_unused]] bool success = chestInventory->mergeItemStackInInventory(copy);
-		}
-
-		// Update the source in the real inventory before re-merging
-		if (slot <= 26) {
-			chestInventory->slots[size_t(slot)] = copy.count == 0 ? ItemStack{} : copy;
-		} else {
-			int playerSlot = slot - 27 + 9;
-			playerInventory->slots[size_t(playerSlot)] = copy.count == 0 ? ItemStack{} : copy;
-		}
-
-		// Re-sync sharedInventory from the real inventories
-		mergeInventories();
-	}
-};
-
-struct PlayerInventoryInteraction : InventoryInteraction {
-	InventoryPlayer* playerInventory;
-
-	PlayerInventoryInteraction(InventoryPlayer* inv) : InventoryInteraction(inv), playerInventory(inv) {}
-
-	virtual bool canExist() override {
-		return playerInventory != nullptr;
-	}
-
-	void onShiftClick(int slot) override {
-		auto from = playerInventory->getInventoryAreaFromSlot(slot);
-		auto stack = playerInventory->getStackInSlot(slot);
-		if (!stack)
-			return;
-
-		ItemStack copy = *stack; // work on a copy so we can detect what moved
-
-		if (from == InvMap::ARMOR || from == InvMap::CRAFTING_AREA || from == InvMap::CRAFTING_RESULT ||
-		    from == InvMap::HOTBAR) {
-			bool success = playerInventory->mergeItemStackInInventory(copy, false, 9, 35);
-			if (!success)
-				playerInventory->mergeItemStackInInventory(copy, false, 36, 44);
-		} else {
-			playerInventory->mergeItemStackInInventory(copy, false, 36, 44);
-		}
-
-		// Update the source slot to whatever count is left
-		if (copy.count == 0) {
-			playerInventory->clearSlot(slot);
-		} else {
-			stack->count = copy.count;
-		}
-	}
+	virtual void onShiftClick(int slot);
 };
